+++ Configuration/Leveling.csv
@@ -1,3 +1,5 @@
 ID,xp,Level,TotalXP
-754894159403286531 & 811828700844654592,67,2,267
+754894159403286531 & 811828700844654592,438,7,2638
 821260011853316166 & 811828700844654592,10,0,10
+510480545160101898 & 811828700844654592,22,4,722
+706839448620498965 & 811828700844654592,41,1,141
